refactor: ajustes de diretório
