+++ src/assets/images/planet-icon.png
