+++ src/assets/images/purple-polygon.png
