+++ src/components/Header/index.js
@@ -1,4 +1,4 @@
-import Logo from "../../assets/logo.png";
+import Logo from "../../assets/images/logo.png";
 import { Container } from "./style";
 
 function Header() {
